+++ src/client/title/title_one.rs
@@ -1,7 +1,7 @@
 use crate::authorization::HEADER_KEY;
+use crate::record::title::TitleRecord;
 use crate::response::error::{ErrorResponse, ErrorResponseBody};
 use crate::response::title::title_one::GetTitleResponse;
-use crate::response::title::TitleRecord;
 use crate::Query;
 
 const RESOURCE: &str = "titles";
