+++ src/record.rs
@@ -1,1 +1,2 @@
+pub mod title;
 pub mod user;
