+++ src/response/title.rs
@@ -1,13 +1,2 @@
-use serde::{Deserialize, Serialize};
-
 pub mod title_one;
 pub mod titles;
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct TitleRecord {
-    pub id: String,
-    pub code: String,
-    pub name: String,
-    pub description: String,
-    pub level: i32,
-}
